requirements.txt
openai
pillow
google-genai
dotenv
pydantic